Put files together and match tensorflow naming convention
