+++ tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag_backward_kernels.cu.cc
@@ -22,8 +22,7 @@ limitations under the License.
 #include <thrust/sort.h>
 
 #include "tensorflow/core/util/gpu_kernel_helper.h"
-#include "tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag.h"
-#include "tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag_backward.h"
+#include "tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag_ops.h"
 
 constexpr int MAX_THREADS_PER_BLOCK = 1024;
