+++ tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag_ops.cc
@@ -19,10 +19,10 @@ limitations under the License.
 #define EIGEN_USE_GPU
 #endif  // GOOGLE_CUDA
 
+#include "tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag_ops.h"
+
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow/core/framework/tensor_shape.h"
-#include "tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag.h"
-#include "tensorflow_addons/custom_ops/layers/cc/kernels/embedding_bag_backward.h"
 
 namespace tensorflow {
 namespace addons {
@@ -31,6 +31,50 @@ typedef Eigen::ThreadPoolDevice CPUDevice;
 typedef Eigen::GpuDevice GPUDevice;
 
 namespace functor {
+// CPU specialization of actual computation.
+template <typename T, typename Tindices>
+struct EmbeddingBagFunctor<CPUDevice, T, Tindices> {
+  static constexpr int64 kPacketSize = Eigen::internal::packet_traits<T>::size;
+  using VectorMap = Eigen::Map<Eigen::Vector<T, Eigen::Dynamic>>;
+  using ConstVectorMap = Eigen::Map<const Eigen::Vector<T, Eigen::Dynamic>>;
+
+  void operator()(const CPUDevice& device,
+                  typename TTypes<Tindices, 2>::ConstTensor indices,
+                  typename TTypes<T, 2>::ConstTensor values,
+                  typename TTypes<T, 2>::ConstTensor weights,
+                  typename TTypes<T, 2>::Tensor output, Combiner combiner) {
+    const Eigen::Index bags = indices.dimension(0);
+    const Eigen::Index sequence_length = indices.dimension(1);
+    const Eigen::Index output_dim = values.dimension(1);
+
+    const auto work = [&](Eigen::Index start, Eigen::Index end) {
+      for (Eigen::Index bag = start; bag < end; ++bag) {
+        VectorMap output_slice(&output(bag, 0), output_dim);
+        output_slice.setZero();
+        for (Eigen::Index seq = 0; seq < sequence_length; ++seq) {
+          const ConstVectorMap values_slice(&values(indices(bag, seq), 0),
+                                            output_dim);
+          output_slice += values_slice * weights(bag, seq);
+        }
+        if (combiner == Combiner::kMean) {
+          output_slice /= static_cast<T>(sequence_length);
+        }
+      }
+    };
+
+    const double bytes_loaded =
+        (sequence_length * output_dim) * (sizeof(Tindices) + 2 * sizeof(T));
+    const double bytes_stored = (sequence_length * output_dim) * sizeof(T);
+    const double compute_cycles =
+        (sequence_length * output_dim) *
+        (Eigen::TensorOpCost::AddCost<T>() + Eigen::TensorOpCost::MulCost<T>());
+    const Eigen::TensorOpCost cost(bytes_loaded, bytes_stored, compute_cycles,
+                                   /*vectorized=*/true,
+                                   /*packet_size=*/kPacketSize);
+    device.parallelFor(bags, cost, std::move(work));
+  }
+};
+
 // CPU specialization of actual computation.
 template <typename T, typename Tindices>
 struct EmbeddingBagBackwardFunctor<CPUDevice, T, Tindices> {
@@ -111,6 +155,50 @@ struct EmbeddingBagBackwardFunctor<CPUDevice, T, Tindices> {
 }  // namespace functor
 
 template <typename Device, typename T, typename Tindices>
+class EmbeddingBagOp : public OpKernel {
+ public:
+  explicit EmbeddingBagOp(OpKernelConstruction* context) : OpKernel(context) {
+    std::string combiner_string;
+    OP_REQUIRES_OK(context, context->GetAttr("combiner", &combiner_string));
+    OP_REQUIRES_OK(context, ValidateCombiner(combiner_string, &combiner_));
+  }
+
+  void Compute(OpKernelContext* context) override {
+    const Tensor& indices = context->input(0);
+    const Tensor& values = context->input(1);
+    const Tensor& weights = context->input(2);
+
+    const TensorShape& indices_shape = indices.shape();
+    const TensorShape& values_shape = values.shape();
+    const TensorShape& weights_shape = weights.shape();
+
+    OP_REQUIRES(
+        context, TensorShapeUtils::IsMatrix(indices_shape),
+        errors::InvalidArgument("indices shape should be at least 2-D."));
+    OP_REQUIRES(context, indices_shape == weights_shape,
+                errors::InvalidArgument(
+                    "Shape of indices and weights should be equal."));
+    OP_REQUIRES(context, TensorShapeUtils::IsMatrix(values_shape),
+                errors::InvalidArgument("values shape should be 2-D."));
+
+    TensorShape output_shape = indices_shape;
+    Eigen::Index output_dim = values.shape().dim_size(1);
+    output_shape.set_dim(output_shape.dims() - 1, output_dim);
+
+    Tensor* output = nullptr;
+    OP_REQUIRES_OK(context, context->allocate_output(0, output_shape, &output));
+
+    functor::EmbeddingBagFunctor<Device, T, Tindices>()(
+        context->eigen_device<Device>(), indices.tensor<Tindices, 2>(),
+        values.tensor<T, 2>(), weights.tensor<T, 2>(), output->tensor<T, 2>(),
+        combiner_);
+  }
+
+ private:
+  Combiner combiner_;
+};
+
+template <typename Device, typename T, typename Tindices>
 class EmbeddingBagBackwardOp : public OpKernel {
  public:
   explicit EmbeddingBagBackwardOp(OpKernelConstruction* context)
@@ -144,7 +232,17 @@ class EmbeddingBagBackwardOp : public OpKernel {
 };
 
 // Register the CPU kernels.
-#define REGISTER_CPU(T)                                                 \
+#define REGISTER_CPU_KERNEL(T)                                          \
+  REGISTER_KERNEL_BUILDER(Name("Addons>EmbeddingBag")                   \
+                              .Device(DEVICE_CPU)                       \
+                              .TypeConstraint<T>("T")                   \
+                              .TypeConstraint<int32>("Tindices"),       \
+                          EmbeddingBagOp<CPUDevice, T, int32>);         \
+  REGISTER_KERNEL_BUILDER(Name("Addons>EmbeddingBag")                   \
+                              .Device(DEVICE_CPU)                       \
+                              .TypeConstraint<T>("T")                   \
+                              .TypeConstraint<int64>("Tindices"),       \
+                          EmbeddingBagOp<CPUDevice, T, int64>);         \
   REGISTER_KERNEL_BUILDER(Name("Addons>EmbeddingBagGrad")               \
                               .Device(DEVICE_CPU)                       \
                               .TypeConstraint<T>("T")                   \
@@ -155,21 +253,28 @@ class EmbeddingBagBackwardOp : public OpKernel {
                               .TypeConstraint<T>("T")                   \
                               .TypeConstraint<int64>("Tindices"),       \
                           EmbeddingBagBackwardOp<CPUDevice, T, int64>);
-REGISTER_CPU(Eigen::half);
-REGISTER_CPU(float);
-REGISTER_CPU(double);
+REGISTER_CPU_KERNEL(Eigen::half);
+REGISTER_CPU_KERNEL(float);
+REGISTER_CPU_KERNEL(double);
+#undef REGISTER_CPU_KERNEL
 
 // Register the GPU kernels.
 #if GOOGLE_CUDA
-// #define REGISTER_GPU(Tindices)                                             \
-//   extern template struct EmbeddingBagBackwardFunctor<GPUDevice, Tindices>; \
-//   REGISTER_KERNEL_BUILDER(Name("Addons>EmbeddingBagGrad")                  \
-//                               .Device(DEVICE_GPU)                          \
-//                               .TypeConstraint<Tindices>("Tindices"),       \
-//                           EmbeddingBagBackwardOp<GPUDevice, Tindices>);
-
-// REGISTER_GPU(int32);
-// REGISTER_GPU(int64);
+#define REGISTER_GPU_KERNEL(T)                                    \
+  REGISTER_KERNEL_BUILDER(Name("Addons>EmbeddingBag")             \
+                              .Device(DEVICE_GPU)                 \
+                              .TypeConstraint<T>("T")             \
+                              .TypeConstraint<int32>("Tindices"), \
+                          EmbeddingBagOp<GPUDevice, T, int32>);   \
+  REGISTER_KERNEL_BUILDER(Name("Addons>EmbeddingBag")             \
+                              .Device(DEVICE_GPU)                 \
+                              .TypeConstraint<T>("T")             \
+                              .TypeConstraint<int64>("Tindices"), \
+                          EmbeddingBagOp<GPUDevice, T, int64>);
+REGISTER_GPU_KERNEL(Eigen::half);
+REGISTER_GPU_KERNEL(float);
+REGISTER_GPU_KERNEL(double);
+#undef REGISTER_GPU_KERNEL
 #endif  // GOOGLE_CUDA
 }  // namespace addons
 }  // namespace tensorflow
